requirements.txt
flask-mysql 
flask-restx 
requests 
pymysql